server/requirements.txt
fastapi~=0.103.1
starlette~=0.27.0
python-dotenv~=1.0.0
pymongo~=4.5.0
uvicorn~=0.22.0
icecream~=2.1.3
requests~=2.26.0
